Rewrite sourcecatalog ingestion for flexibility
Deleted precovery.sourcecatalog.SourceExposure. Its functionality was
purely internal to the sourcecatalog module, and it can be
accomplished using a dictionary instead, which is less code weight.

Rewrote most of sourcecatalog. Instead of primarily operating on CSV
files, it breaks that step into two phases. One phase is loading in a
stream of SourceObservations. This could be from a CSV file, but could
also be from a DataFrame, for example. The second phase is organizing
the SourceObservations into chunks grouped by HEALPixel
("SourceFrames"), suitable for loading into the database. Those two
phases are now contained in separate functions.